Allow function params to be arrays (decay to pointers)
Also, allow cast to void, and make it so that it can't be recast to
anything else. The value is essentially dead.

TODO: do dead-code elim pass to eliminate any loads of the type like
`(void)0`, no side-effects in expression `0`, so safe to eliminate
it.

+++ mine/test/compiler/argument_reg_noclobber.c
@@ -5,9 +5,9 @@ int mycall(int a, int b, int c, int d, int e) {
 }
 
 int main(void) {
-  long i;
+  int i;
   i = 10;
-  printf("%ld\n", i);
+  printf("%d\n", i);
   /**
    * mycall() shouldn't clobber argument register 4 for printf() (rcx)
    * In order to do this, for a call to a function we make a first-pass thru
@@ -20,6 +20,6 @@ int main(void) {
    * Brute force: for every function call cgloadarg, we push all argument
    * values to the stack before the call and pop them off after the call.
    */
-  printf("%ld, %d, %ld\n", i, mycall(1, 2, 3, 4, 5), i);
+  printf("%d, %d, %d\n", i, mycall(1, 2, 3, 4, 5), i);
   return (0);
 }
